requirements.txt
django==4.2.2
djangorestframework==3.14.0
psycopg2
djangorestframework-simplejwt
drf-yasg
environs==9.5.0
cryptography